Add an option to duplicate a cabin and logos for duplicating, editing and deleting

+++ src/services/apiCabins.js
@@ -40,10 +40,15 @@ export async function createEditCabin(newCabin, id) {
 		throw new Error('Cabin could not be created');
 	}
 
+	// upload image
+
+	if (hasImagePath) return data;
+
 	const { error: storageError } = await supabase.storage
 		.from('cabin-images')
 		.upload(imageName, newCabin.image);
 
+	// delete cabin if there was an error uploading image
 	if (storageError) {
 		await supabase.from('cabins').delete().eq('id', data.id);
 		console.error(storageError);
